+++ .storybook/preview.ts
@@ -10,6 +10,8 @@ import '../packages/ui/dist/css/reset.css';
 import '../packages/ui/dist/css/components/button.css';
 import '../packages/ui/dist/css/components/container.css';
 import '../packages/ui/dist/css/components/link.css';
+import '../packages/ui/dist/css/components/label-standalone.css';
+import '../packages/ui/dist/css/components/label-wrapper.css';
 import '../packages/ui/dist/css/components/split-surface.css';
 import '../packages/ui/dist/css/components/text.css';
 import '../packages/ui/dist/css/components/body.css';