+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @rei/c1-ui Changelog
 
+## 0.1.11-alpha.0 - 2026-01-31
+
+- Added label standalone and label wrapper component styles.
+
 ## 0.1.10-alpha.0 - 2026-01-31
 
 - Added title component styles with container query scaling.